requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
Pillow==9.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.2

channels~=3.0.5
djangorestframework~=3.13.1